+++ tests/test_two_periods_multiple_exog.py
@@ -4,9 +4,8 @@ import jax.numpy as jnp
 import numpy as np
 import pandas as pd
 import pytest
-from dcegm.pre_processing.process_model import (
-    process_model_functions_and_create_state_space_objects,
-)
+from dcegm.pre_processing.model_functions import process_model_functions
+from dcegm.pre_processing.state_space import create_state_space_and_choice_objects
 from dcegm.solve import solve_dcegm
 from numpy.testing import assert_allclose
 from scipy.special import roots_sh_legendre
@@ -19,18 +18,22 @@ from toy_models.consumption_retirement_model.state_space_objects import (
 )
 from toy_models.consumption_retirement_model.state_space_objects import update_state
 
-from tests.two_period_models.ltc_and_job_offer.dcegm_code import (
+from tests.two_period_models.exog_ltc_and_job_offer.euler_equation import (
+    euler_rhs_two_exog_processes,
+)
+from tests.two_period_models.exog_ltc_and_job_offer.model_functions import (
     budget_dcegm_two_exog_processes,
 )
-from tests.two_period_models.ltc_and_job_offer.dcegm_code import flow_util
-from tests.two_period_models.ltc_and_job_offer.dcegm_code import func_exog_job_offer
-from tests.two_period_models.ltc_and_job_offer.dcegm_code import func_exog_ltc
-from tests.two_period_models.ltc_and_job_offer.dcegm_code import (
+from tests.two_period_models.exog_ltc_and_job_offer.model_functions import flow_util
+from tests.two_period_models.exog_ltc_and_job_offer.model_functions import (
+    func_exog_job_offer,
+)
+from tests.two_period_models.exog_ltc_and_job_offer.model_functions import func_exog_ltc
+from tests.two_period_models.exog_ltc_and_job_offer.model_functions import (
     inverse_marginal_utility,
 )
-from tests.two_period_models.ltc_and_job_offer.dcegm_code import marginal_utility
-from tests.two_period_models.ltc_and_job_offer.eueler_equation_code import (
-    euler_rhs_two_exog_processes,
+from tests.two_period_models.exog_ltc_and_job_offer.model_functions import (
+    marginal_utility,
 )
 
 WEALTH_GRID_POINTS = 100
@@ -60,7 +63,7 @@ def utility_functions():
 
 
 @pytest.fixture(scope="module")
-def input_data_two_exog_processes(state_space_functions, utility_functions):
+def input_data(state_space_functions, utility_functions):
     # ToDo: Write this as dictionary such that it has a much nicer overview
     index = pd.MultiIndex.from_tuples(
         [("utility_function", "rho"), ("utility_function", "delta")],
@@ -132,7 +135,7 @@ def input_data_two_exog_processes(state_space_functions, utility_functions):
     TEST_CASES_TWO_EXOG_PROCESSES,
 )
 def test_two_period_two_exog_processes(
-    input_data_two_exog_processes,
+    input_data,
     wealth_idx,
     state_idx,
     utility_functions,
@@ -141,25 +144,42 @@ def test_two_period_two_exog_processes(
     quad_points, quad_weights = roots_sh_legendre(5)
     quad_draws = norm.ppf(quad_points) * 1
 
-    params = input_data_two_exog_processes["params"]
+    params = input_data["params"]
     keys = params.index.droplevel("category").tolist()
     values = params["value"].tolist()
     params = dict(zip(keys, values))
+
     (
-        *_,
-        period_specific_state_objects,
-        state_space,
-    ) = process_model_functions_and_create_state_space_objects(
-        options=input_data_two_exog_processes["options"],
+        compute_utility,
+        compute_marginal_utility,
+        compute_inverse_marginal_utility,
+        compute_beginning_of_period_wealth,
+        compute_final_period,
+        compute_exog_transition_vec,
+        compute_upper_envelope,
+        get_state_specific_choice_set,
+        update_endog_state_by_state_and_choice,
+    ) = process_model_functions(
+        input_data["options"],
         user_utility_functions=utility_functions,
         user_budget_constraint=budget_dcegm_two_exog_processes,
         user_final_period_solution=solve_final_period_scalar,
         state_space_functions=state_space_functions,
     )
+
+    (
+        period_specific_state_objects,
+        state_space,
+    ) = create_state_space_and_choice_objects(
+        options=input_data["options"],
+        get_state_specific_choice_set=get_state_specific_choice_set,
+        update_endog_state_by_state_and_choice=update_endog_state_by_state_and_choice,
+    )
+
     period = state_space["period"][state_idx]
 
-    endog_grid_period = input_data_two_exog_processes["result"][period]["endog_grid"]
-    policy_period = input_data_two_exog_processes["result"][period]["policy_left"]
+    endog_grid_period = input_data["result"][period]["endog_grid"]
+    policy_period = input_data["result"][period]["policy_left"]
 
     state_choices_period = period_specific_state_objects[period]["state_choice_mat"]
