+++ tests/two_period_models/exog_ltc_and_job_offer/model_functions.py
